+++ source/internal/core/class/Instance.js
